Avertissements masqués

+++ InverseWayLiberation/src/Physics/Bodies/DynamicBox.cpp
@@ -29,7 +29,7 @@ DynamicBox::DynamicBox(World *world, b2Vec3 posRot, std::shared_ptr<sf::Texture>
 		fixtureDef.density = density;
 		fixtureDef.friction = friction;
 		fixtureDef.restitution = restitution;
-		fixtureDef.filter.groupIndex = groupIndex;
+		fixtureDef.filter.groupIndex = static_cast<int16>(groupIndex);
 		fixtureDef.filter.categoryBits = categoryBits;
 		fixtureDef.filter.maskBits = maskBits;
 		fixtureDef.shape = mShape;
@@ -67,7 +67,7 @@ DynamicBox::DynamicBox(World *world, b2Vec3 posRot, sf::Sprite *sprite, float de
 		fixtureDef.density = density;
 		fixtureDef.friction = friction;
 		fixtureDef.restitution = restitution;
-		fixtureDef.filter.groupIndex = groupIndex;
+		fixtureDef.filter.groupIndex = static_cast<int16>(groupIndex);
 		fixtureDef.filter.categoryBits = categoryBits;
 		fixtureDef.filter.maskBits = maskBits;
 		fixtureDef.shape = mShape;
