+++ InverseWayLiberation/src/Physics/Bodies/DynamicCircle.cpp
@@ -28,7 +28,7 @@ DynamicCircle::DynamicCircle(World *world, b2Vec3 posRot, std::shared_ptr<sf::Te
 		fixtureDef.density = density;
 		fixtureDef.friction = friction;
 		fixtureDef.restitution = restitution;
-		fixtureDef.filter.groupIndex = groupIndex;
+		fixtureDef.filter.groupIndex = static_cast<int16>(groupIndex);
 		fixtureDef.filter.categoryBits = categoryBits;
 		fixtureDef.filter.maskBits = maskBits;
 		fixtureDef.shape = mShape;
